+++ app/src/androidTest/java/com/github/sdp/ratemyepfl/database/FakeClassroomsReviewsRepository.kt
@@ -1,6 +1,5 @@
 package com.github.sdp.ratemyepfl.database
 
-import com.github.sdp.ratemyepfl.model.review.ClassroomReview
 import java.time.LocalDate
 import javax.inject.Inject
