+++ app/src/androidTest/java/com/github/sdp/ratemyepfl/database/FakeCoursesReviewsRepository.kt
@@ -1,7 +1,5 @@
 package com.github.sdp.ratemyepfl.database
 
-import com.github.sdp.ratemyepfl.model.review.CourseReview
-import com.github.sdp.ratemyepfl.model.review.Review
 import com.github.sdp.ratemyepfl.model.review.ReviewRating
 import java.time.LocalDate
 import javax.inject.Inject
