+++ app/src/main/java/com/github/sdp/ratemyepfl/adapter/CourseReviewAdapter.kt
@@ -7,7 +7,6 @@ import android.view.ViewGroup
 import android.widget.ArrayAdapter
 import android.widget.TextView
 import com.github.sdp.ratemyepfl.R
-import com.github.sdp.ratemyepfl.model.review.CourseReview
 
 class CourseReviewAdapter(
     context: Context,
