+++ app/src/main/java/com/github/sdp/ratemyepfl/adapter/RoomReviewsAdapter.kt
@@ -7,7 +7,6 @@ import android.widget.TextView
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
-import com.github.sdp.ratemyepfl.model.review.ClassroomReview
 import com.github.sdp.ratemyepfl.R
 
 class RoomReviewsAdapter :
