+++ app/src/main/java/com/github/sdp/ratemyepfl/database/ClassroomsReviewsRepositoryInterface.kt
@@ -1,7 +1,5 @@
 package com.github.sdp.ratemyepfl.database
 
-import com.github.sdp.ratemyepfl.model.review.ClassroomReview
-
 interface ClassroomsReviewsRepositoryInterface {
     fun add(value: ClassroomReview)
     suspend fun get(): List<ClassroomReview?>
