+++ app/src/main/java/com/github/sdp/ratemyepfl/database/CoursesReviewsRepositoryInterface.kt
@@ -1,7 +1,5 @@
 package com.github.sdp.ratemyepfl.database
 
-import com.github.sdp.ratemyepfl.model.review.CourseReview
-
 interface CoursesReviewsRepositoryInterface {
     suspend fun get() : List<CourseReview?>
     fun add(value: CourseReview)
