+++ app/src/main/java/com/github/sdp/ratemyepfl/model/items/Classroom.kt
@@ -1,6 +1,5 @@
 package com.github.sdp.ratemyepfl.model.items
 
-import com.github.sdp.ratemyepfl.model.review.ClassroomReview
 import com.google.firebase.firestore.DocumentSnapshot
 import kotlinx.serialization.Serializable
