+++ app/src/main/java/com/github/sdp/ratemyepfl/viewmodel/CourseReviewListViewModel.kt
@@ -4,8 +4,6 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.github.sdp.ratemyepfl.model.review.CourseReview
-import com.github.sdp.ratemyepfl.database.CoursesReviewsRepository
 import com.github.sdp.ratemyepfl.database.CoursesReviewsRepositoryInterface
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
